feat: refactor frontend/
