+++ frontend/src/DogDetailsPage/ImagesSlider.tsx
