+++ frontend/src/DogDetailsPage/VideosSlider.tsx
