+++ frontend/src/helper_functions.tsx
